+++ Changelog.md
@@ -1,6 +1,13 @@
 ### 3.0.0.rc1 Development
 [Full Changelog](http://github.com/rspec/rspec-expectations/compare/v3.0.0.beta2...master)
 
+Breaking Changes:
+
+* Remove `matcher_execution_context` attribute from DSL-defined
+  custom matchers. (Myron Marston)
+* Remove `RSpec::Matchers::Pretty#_pretty_print`. (Myron Marston)
+* Remove `RSpec::Matchers::Pretty#expected_to_sentence`. (Myron Marston)
+
 Bug Fixes:
 
 * Fix `NoMethodError` triggered by beta2 when `YARD` was loaded in